+++ docs/changelog.md
@@ -1,5 +1,6 @@
 ### Feb 06, 2018 - Version 3.8.3
 - Added: Option for submit button to print or save PDF based on custom HTML that supports {tags} to dynamically retrieve form data
+- Added: Print button can support signatures when used like `<embed type="image/png" src="{signature}"></embed>`
 
 ### Jan 29, 2018 - Version 3.8.0
 - Added: Compatibility for variable fields with advanced tags e.g: {field;2}